new group.cpp
